updated graph evaluation

+++ evaluation/RealBank/04_evaluation_RebuildGraph.py
@@ -11,11 +11,6 @@ from scipy.spatial.distance import pdist, cdist
 from sklearn.cluster import KMeans, MiniBatchKMeans
 from sklearn.preprocessing import StandardScaler
 
-import networkx as nx
-
-from modified_sitepackages.netcomp import netsimile, deltacon0
-
-
 models = ['DOPPELGANGER', 'FINDIFF', 'TVAE', 'WGANGPwDRS', 'CTGAN']
 keep_col = ['PC1','PC2','PC3','PC4','PC5','PC6','PC7','PC8','PC9','PC10']
 
@@ -67,45 +62,51 @@ for model in models:
     synth_nodes_avg_distances = {}
     synthetic_nodes_scaled = pd.DataFrame(scaler.fit_transform(synthetic_nodes), columns=synthetic_nodes.columns)
 
-    for n_clusters in tqdm(range(int(num_real_nodes*0.25), int(num_real_nodes*1.5), 30000), desc="Search for optimal Number of Nodes"):
-        if not os.path.exists("./working/kmeans_{}/kmeans_{}_{}.pkl".format(model, model, n_clusters)):
-            kms = MiniBatchKMeans(n_clusters= n_clusters, init= "k-means++", n_init= "auto", batch_size= 8192)
-            kms.fit(synthetic_nodes_scaled)
-            if not os.path.exists("./working/kmeans_{}".format(model)):
-                os.makedirs("./working/kmeans_{}".format(model))
-            with open("./working/kmeans_{}/kmeans_{}_{}.pkl".format(model, model, n_clusters), 'wb') as f:
-                pickle.dump(kms, f)
-        else:
-            with open("./working/kmeans_{}/kmeans_{}_{}.pkl".format(model, model, n_clusters), 'rb') as f:
-                kms = pickle.load(f)
+    if not os.path.exists("./synth/{}_synthetic_data_graph_distance.csv".format(model)):
+        for n_clusters in tqdm(range(int(num_real_nodes*0.25), int(num_real_nodes*1.5), 30000), desc="Search for optimal Number of Nodes"):
+            if not os.path.exists("./working/kmeans_{}/kmeans_{}_{}.pkl".format(model, model, n_clusters)):
+                kms = MiniBatchKMeans(n_clusters= n_clusters, init= "k-means++", n_init= "auto", batch_size= 8192)
+                kms.fit(synthetic_nodes_scaled)
+                if not os.path.exists("./working/kmeans_{}".format(model)):
+                    os.makedirs("./working/kmeans_{}".format(model))
+                with open("./working/kmeans_{}/kmeans_{}_{}.pkl".format(model, model, n_clusters), 'wb') as f:
+                    pickle.dump(kms, f)
+            else:
+                with open("./working/kmeans_{}/kmeans_{}_{}.pkl".format(model, model, n_clusters), 'rb') as f:
+                    kms = pickle.load(f)
 
-        new_synthetic_nodes = pd.DataFrame(kms.cluster_centers_)
+            new_synthetic_nodes = pd.DataFrame(kms.cluster_centers_)
 
-        synthetic_nodes_avg_distance = []
-        for i, chunk in new_synthetic_nodes.groupby(np.arange(len(new_synthetic_nodes)) // 15000):
-            synthetic_nodes_avg_distance.append(np.mean(cdist(new_synthetic_nodes.drop(chunk.index).values, chunk.values, 'euclid')))
-        synthetic_nodes_avg_distance = np.mean(synthetic_nodes_avg_distance)
+            synthetic_nodes_avg_distance = []
+            for i, chunk in new_synthetic_nodes.groupby(np.arange(len(new_synthetic_nodes)) // 15000):
+                synthetic_nodes_avg_distance.append(np.mean(cdist(new_synthetic_nodes.drop(chunk.index).values, chunk.values, 'euclid')))
+            synthetic_nodes_avg_distance = np.mean(synthetic_nodes_avg_distance)
 
-        synth_nodes_avg_distances[n_clusters] = synthetic_nodes_avg_distance
+            synth_nodes_avg_distances[n_clusters] = synthetic_nodes_avg_distance
 
-    n_clusters = min(synth_nodes_avg_distances, key=lambda x:abs(x- real_nodes_avg_distance))
-    with open("./working/kmeans_{}/kmeans_{}_{}.pkl".format(model, model, n_clusters), 'rb') as f:
-        kms = pickle.load(f)
-    synthetic_nodes_distance = synthetic_nodes
-    synthetic_nodes_distance["node_id"] = kms.predict(synthetic_nodes_distance)
+        n_clusters = min(synth_nodes_avg_distances, key=lambda x:abs(x- real_nodes_avg_distance))
+        with open("./working/kmeans_{}/kmeans_{}_{}.pkl".format(model, model, n_clusters), 'rb') as f:
+            kms = pickle.load(f)
+        synthetic_nodes_distance = synthetic_nodes
+        synthetic_nodes_distance["node_id"] = kms.predict(synthetic_nodes_distance)
 
-    synthetic_data_distance = synthetic_data.merge(synthetic_nodes_distance, left_on= ["target_id_{}".format(i) for i in range(6)], right_on= ["id_{}".format(i) for i in range(6)], how="left", suffixes= ("", "_target"))
-    synthetic_data_distance = synthetic_data_distance.merge(synthetic_nodes_distance, left_on=["source_id_{}".format(i) for i in range(6)], right_on=["id_{}".format(i) for i in range(6)], how="left", suffixes= ("", "_source"))
+        synthetic_data_distance = synthetic_data.merge(synthetic_nodes_distance, left_on= ["target_id_{}".format(i) for i in range(6)], right_on= ["id_{}".format(i) for i in range(6)], how="left", suffixes= ("", "_target"))
+        synthetic_data_distance = synthetic_data_distance.merge(synthetic_nodes_distance, left_on=["source_id_{}".format(i) for i in range(6)], right_on=["id_{}".format(i) for i in range(6)], how="left", suffixes= ("", "_source"))
 
-    synthetic_data_distance.to_csv("./synth/{}_synthetic_data_graph_distance.csv".format(model), index=False)
+        synthetic_data_distance.to_csv("./synth/{}_synthetic_data_graph_distance.csv".format(model), index=False)
 
-    n_clusters = int(real_nodes_per_datapoint * synthetic_data.shape[0])
-    kms = MiniBatchKMeans(n_clusters=n_clusters, init="k-means++", n_init="auto", batch_size=8192)
-    kms.fit(synthetic_nodes_scaled)
-    synthetic_nodes_number = synthetic_nodes
-    synthetic_nodes_number["node_id"] = kms.predict(synthetic_nodes_number)
+    if not os.path.exists("./synth/{}_synthetic_data_graph_number.csv".format(model)):
+        n_clusters = int(real_nodes_per_datapoint * synthetic_data.shape[0])
+        kms = MiniBatchKMeans(n_clusters=n_clusters, init="k-means++", n_init="auto", batch_size=8192)
+        kms.fit(synthetic_nodes_scaled)
+        synthetic_nodes_number = synthetic_nodes
+        synthetic_nodes_number["node_id"] = kms.predict(synthetic_nodes_number)
 
-    synthetic_data_graph_number = synthetic_data.merge(synthetic_nodes_number, left_on=["target_id_{}".format(i) for i in range(6)], right_on=["id_{}".format(i) for i in range(6)], how="left", suffixes=("", "_target"))
-    synthetic_data_graph_number = synthetic_data_graph_number.merge(synthetic_nodes_number, left_on=["source_id_{}".format(i) for i in range(6)], right_on=["id_{}".format(i) for i in range(6)], how="left", suffixes=("", "_source"))
+        if "source_id" in synthetic_data.columns:
+            synthetic_data = synthetic_data.drop(columns= ["source_id"])
+        synthetic_nodes_number = synthetic_nodes_number.rename(columns= dict(zip(["id_{}".format(i) for i in range(6)] + ["node_id"], ["target_id_{}".format(i) for i in range(6)] + ["target_id"])))
+        synthetic_data_graph_number = synthetic_data.merge(synthetic_nodes_number, on=["target_id_{}".format(i) for i in range(6)], how="left", suffixes=("", "_target"))
+        synthetic_nodes_number = synthetic_nodes_number.rename(columns=dict(zip(["target_id_{}".format(i) for i in range(6)] + ["target_id"], ["source_id_{}".format(i) for i in range(6)] + ["source_id"])))
+        synthetic_data_graph_number = synthetic_data_graph_number.merge(synthetic_nodes_number, on=["source_id_{}".format(i) for i in range(6)], how="left", suffixes=("", "_source"))
 
-    synthetic_data_graph_number.to_csv("./synth/{}_synthetic_data_graph_number.csv".format(model), index=False)
+        synthetic_data_graph_number.to_csv("./synth/{}_synthetic_data_graph_number.csv".format(model), index=False)
